--- src/model/otp.ts
@@ -1,5 +1,0 @@
-export type OTP = {
-    code: string,
-    sentAt: number,
-    validatedAt?: number;
-}
